[noissue] renamed "lib" to "src"
