+++ src/docuWriter/entityValidator.js
