+++ src/docuWriter/identifierSanitizer.js
