+++ src/docuWriter/pageNameExtractor.js
