+++ src/docuWriter/util.js
