+++ src/docuWriter/validationSchemas/details.json
