+++ src/docuWriter/validationSchemas/scenario.json
