+++ src/scenariooReporter.js
