+++ src/scenariooStore.js
